Add crash standalone version check and GPG signature verification
While we currently sign and verify our CrateDB tarballs, we do not do
this with our Crash standalone releases. Now that we are downloading and
including them in our Docker build phase, we ought to do this too.

This PR therefore verifies the Crash standalone downloads via
downloading and checking the GPG signature file, in the same manner as
the tarball.

This inclusion of the GPG signature for crash standalones should be a
part of the crash release process from 0.24.2 onwards.