Base Crate Docker image from CentOS 7, not Alpine
The purpose of this commit is to fix issues that we have encountered
with Docker Swarm, caused by problems with Java/libc complications
within Alpine Linux.

This commit, firstly, changes the base image from Alpine 3.7 to CentOS 7
and installs the various components we need via yum, instead of apk.

It also removes the use of the gosu tool, instead relying on chroot.
This should also let people build the docker image for ARM, as a side
effect.

It also uses Jinja2 templating as a templating engine, rather than using
sed to search-and-replace text strings.

It also adds some metadata labels for the Docker image, as described in
best practices. Though this isn't required for this shift, it seemed
sensible to include.

The original Alpine image's size was 198MB. The size of this resulting
image is 475MB.