Remove installing the IUS repository
IUS (Inline with Upstream Stable) is a yum repository with newer versions of
some software for Centos. However, when adding this repo on the ARM architecture,
we get the following error:

https://repo.ius.io/7/aarch64/repodata/repomd.xml: [Errno 14] HTTPS Error 404 - Not Found

Tthe IUS repo has no repodata for aarch64 packages, which would suggest that it isn't
supported on that architecture.

The one package we were relying on the IUS repository for was python36u. However, the
standard yum repositories have a python36 package, and both of these packages install
the same python version (3.6.8).

This commit therefore removes the dependency on python36u, switching it to the standard
python36 repository and removing the need to add the IUS repo. This now builds as expected
on linux/arm64.