Fix yum within CrateDB's docker image
Yum, the CentOS package manager, requires Python 2. In the course
of installing Python 3.6, we symlink the default system Python from
Python 2 to 3.

The shebang at the top of yum's main file is '!#/usr/bin/python',
which now references the 3.6 installation. This commit replaces
the shebang at the top of yum's script to point, instead, to
'!#/usr/bin/python2'.