Switch back to almalinux:9 as base image
We discovered regressions when running the AlmaLinux minimal image on
Kubernetes. As a safety measure to not cause any similar problems to
users, we are going back to the AlmaLinux standard image.

When looking at the size savings, we also found that the *compressed*
image size delta compared between standard and minimal is only 27 MB.